refactor: Reorganize/move specific code into components/ dir
