+++ crates/sherwood/src/lib.rs
@@ -1,15 +1,9 @@
-pub mod config;
+pub mod components;
 pub mod generator;
-pub mod html_renderer;
-pub mod markdown_parser;
-pub mod page_generator;
 pub mod server;
-pub mod templates;
-pub mod themes;
-pub mod utils;
 
-pub use config::{SiteConfig, SiteSection, TemplateSection};
+pub use components::config::{SiteConfig, SiteSection, TemplateSection};
+pub use components::templates::{TemplateManager, validate_templates};
+pub use components::themes::{Theme, ThemeManager};
 pub use generator::generate_site;
 pub use server::run_dev_server;
-pub use templates::{TemplateManager, validate_templates};
-pub use themes::{Theme, ThemeManager};
